Memorize the hidden calendars on toggleSchedules and reuse this in createSchedules
Calendar.toggleSchedules() iterates all loaded schedules and switches
some of them on or off.  Schedules loaded later with Calendar.createSchedules()
are displayed, even if the calendar was switched previously off with
Calendar.toggleSchedules().

This patch:
- memorizes in toggleSchedules() whether a calendar is shown or hidden (when
_isVisible is undefined/missing or true, then the calendar is shown)
- in createSchedules() obtains from the calendar information, whether the
latter is visible, and modifies the isVisible property of the schedules
accordingly.

As TUI.calendar runs on IE9+ and the other recent version of browsers, it
does support Array.prototype.forEach.  Replace util.forEach() with the former.