Update README.md
Small and easy to make typos I noticed and fixed. The only bit I'm not certain about is in line 287, "Before PR, first check to test updates then check for any errors." I originally had check "for" test updates but changed it to it "to." I don't know if this is referring to the test cases in 271 but the wording was confusing, at least to me. Anyway, feel free to accept or decline the changes!!! 👍🏻  :) 
 
Thank you so much for an amazing open source project! 

Happy Coding!! 
-Bridgette